feat(chatmessage): type move from chatbot
