+++ src/chat-message/content/thinking-content.tsx
@@ -10,8 +10,8 @@ import { Component, signal, tag } from 'omi';
 import { getClassPrefix } from '../../_util/classname';
 import { convertToLightDomNode } from '../../_util/lightDom';
 import type { ChatMessageStatus } from '../../chat-engine';
-import type { TdChatContentProps } from '../../chatbot';
 import { CollapseValue } from '../../collapse';
+import type { TdChatContentProps } from '../type';
 
 import styles from '../style/chat-item.less';
